(BSR)[API] feat: move all methods related to product creation/update in dedicated `product_service.py`
