+++ api/src/pcapi/routes/public/individual_offers/v1/products.py
@@ -1,28 +1,22 @@
 import copy
-import datetime
 import logging
 
 from flask import request
 import sqlalchemy as sqla
 
 from pcapi import repository
-from pcapi.core import search
-from pcapi.core.categories import subcategories_v2 as subcategories
 from pcapi.core.categories.categories import TITELIVE_MUSIC_TYPES
 from pcapi.core.finance import utils as finance_utils
 from pcapi.core.offerers import api as offerers_api
-from pcapi.core.offerers import models as offerers_models
 from pcapi.core.offers import api as offers_api
 from pcapi.core.offers import exceptions as offers_exceptions
 from pcapi.core.offers import models as offers_models
 from pcapi.core.offers import validation as offers_validation
-from pcapi.core.providers import models as providers_models
 from pcapi.core.providers.constants import TITELIVE_MUSIC_GENRES_BY_GTL_ID
 from pcapi.domain import music_types
 from pcapi.domain import show_types
 from pcapi.models import api_errors
 from pcapi.models import db
-from pcapi.models.offer_mixin import OfferValidationType
 from pcapi.routes.public import spectree_schemas
 from pcapi.routes.public.documentation_constants import http_responses
 from pcapi.routes.public.documentation_constants import tags
@@ -32,13 +26,12 @@ from pcapi.serialization.spec_tree import ExtendResponse as SpectreeResponse
 from pcapi.utils import image_conversion
 from pcapi.validation.routes.users_authentifications import api_key_required
 from pcapi.validation.routes.users_authentifications import current_api_key
-from pcapi.workers import worker
-from pcapi.workers.decorators import job
 
 from . import blueprint
 from . import constants
 from . import serialization
 from . import utils
+from .services import product_service
 
 
 logger = logging.getLogger(__name__)
@@ -284,211 +277,8 @@ def post_product_offer_by_ean(body: serialization.ProductsOfferByEanCreation) ->
     venue = utils.retrieve_venue_from_location(body.location)
     if venue.isVirtual:
         raise api_errors.ApiErrors({"location": ["Cannot create product offer for virtual venues"]})
-    serialized_products_stocks = _serialize_products_from_body(body.products)
-    _create_or_update_ean_offers.delay(serialized_products_stocks, venue.id, current_api_key.provider.id)
-
-
-@job(worker.low_queue)
-def _create_or_update_ean_offers(serialized_products_stocks: dict, venue_id: int, provider_id: int) -> None:
-    provider = providers_models.Provider.query.filter_by(id=provider_id).one()
-    venue = offerers_models.Venue.query.filter_by(id=venue_id).one()
-
-    ean_to_create_or_update = set(serialized_products_stocks.keys())
-
-    offers_to_update = _get_existing_offers(ean_to_create_or_update, venue)
-
-    offer_to_update_by_ean = {}
-    ean_list_to_update = set()
-    for offer in offers_to_update:
-        ean_list_to_update.add(offer.extraData["ean"])  # type: ignore[index]
-        offer_to_update_by_ean[offer.extraData["ean"]] = offer  # type: ignore[index]
-
-    ean_list_to_create = ean_to_create_or_update - ean_list_to_update
-    offers_to_index = []
-    with repository.transaction():
-        if ean_list_to_create:
-            created_offers = []
-            existing_products = _get_existing_products(ean_list_to_create)
-            product_by_ean = {product.extraData["ean"]: product for product in existing_products}  # type: ignore[index]
-            not_found_eans = [ean for ean in ean_list_to_create if ean not in product_by_ean.keys()]
-            if not_found_eans:
-                logger.warning(
-                    "Some provided eans were not found",
-                    extra={"eans": ",".join(not_found_eans), "venue": venue_id},
-                    technical_message_id="ean.not_found",
-                )
-            for product in existing_products:
-                try:
-                    ean = product.extraData["ean"] if product.extraData else None
-                    stock_data = serialized_products_stocks[ean]
-                    created_offer = _create_offer_from_product(
-                        venue,
-                        product_by_ean[ean],
-                        provider,
-                    )
-                    created_offers.append(created_offer)
-
-                except (
-                    offers_exceptions.OfferCreationBaseException,
-                    offers_exceptions.OfferEditionBaseException,
-                ) as exc:
-                    logger.info(
-                        "Error while creating offer by ean",
-                        extra={
-                            "ean": ean,
-                            "venue_id": venue_id,
-                            "provider_id": provider_id,
-                            "exc": exc.__class__.__name__,
-                        },
-                    )
-
-            db.session.bulk_save_objects(created_offers)
-
-            reloaded_offers = _get_existing_offers(ean_list_to_create, venue)
-            for offer in reloaded_offers:
-                try:
-                    ean = offer.extraData["ean"]  # type: ignore[index]
-                    stock_data = serialized_products_stocks[ean]
-                    # FIXME (mageoffray, 2023-05-26): stock saving optimisation
-                    # Stocks are inserted one by one for now, we need to improve create_stock to remove the repository.session.add()
-                    # It will be done before the release of this API
-                    offers_api.create_stock(
-                        offer=offer,
-                        price=finance_utils.to_euros(stock_data["price"]),
-                        quantity=serialization.deserialize_quantity(stock_data["quantity"]),
-                        booking_limit_datetime=stock_data["booking_limit_datetime"],
-                        creating_provider=provider,
-                    )
-                except (
-                    offers_exceptions.OfferCreationBaseException,
-                    offers_exceptions.OfferEditionBaseException,
-                ) as exc:
-                    logger.info(
-                        "Error while creating offer by ean",
-                        extra={
-                            "ean": ean,
-                            "venue_id": venue_id,
-                            "provider_id": provider_id,
-                            "exc": exc.__class__.__name__,
-                        },
-                    )
-
-        for offer in offers_to_update:
-            try:
-                offer.lastProvider = provider
-                offer.isActive = True
-
-                ean = offer.extraData["ean"]  # type: ignore[index]
-                stock_data = serialized_products_stocks[ean]
-                # FIXME (mageoffray, 2023-05-26): stock upserting optimisation
-                # Stocks are edited one by one for now, we need to improve edit_stock to remove the repository.session.add()
-                # It will be done before the release of this API
-                _upsert_product_stock(
-                    offer_to_update_by_ean[ean],
-                    serialization.StockEdition(
-                        **{
-                            "price": stock_data["price"],
-                            "quantity": stock_data["quantity"],
-                            "booking_limit_datetime": stock_data["booking_limit_datetime"],
-                        }
-                    ),
-                    provider,
-                )
-                offers_to_index.append(offer_to_update_by_ean[ean].id)
-            except (offers_exceptions.OfferCreationBaseException, offers_exceptions.OfferEditionBaseException) as exc:
-                logger.info(
-                    "Error while creating offer by ean",
-                    extra={"ean": ean, "venue_id": venue_id, "provider_id": provider_id, "exc": exc.__class__.__name__},
-                )
-
-    search.async_index_offer_ids(
-        offers_to_index,
-        reason=search.IndexationReason.OFFER_UPDATE,
-        log_extra={"venue_id": venue_id, "source": "offers_public_api"},
-    )
-
-
-def _get_existing_products(ean_to_create: set[str]) -> list[offers_models.Product]:
-    allowed_product_subcategories = [
-        subcategories.SUPPORT_PHYSIQUE_MUSIQUE_CD.id,
-        subcategories.SUPPORT_PHYSIQUE_MUSIQUE_VINYLE.id,
-        subcategories.LIVRE_PAPIER.id,
-    ]
-    return offers_models.Product.query.filter(
-        offers_models.Product.extraData["ean"].astext.in_(ean_to_create),
-        offers_models.Product.can_be_synchronized == True,
-        offers_models.Product.subcategoryId.in_(allowed_product_subcategories),
-        # FIXME (cepehang, 2023-09-21) remove these condition when the product table is cleaned up
-        offers_models.Product.lastProviderId.is_not(None),
-        offers_models.Product.idAtProviders.is_not(None),
-    ).all()
-
-
-def _get_existing_offers(
-    ean_to_create_or_update: set[str],
-    venue: offerers_models.Venue,
-) -> list[offers_models.Offer]:
-    subquery = (
-        db.session.query(
-            sqla.func.max(offers_models.Offer.id).label("max_id"),
-        )
-        .filter(offers_models.Offer.isEvent == False)
-        .filter(offers_models.Offer.venue == venue)
-        .filter(offers_models.Offer.extraData["ean"].astext.in_(ean_to_create_or_update))
-        .group_by(offers_models.Offer.extraData["ean"], offers_models.Offer.venueId)
-        .subquery()
-    )
-
-    return (
-        utils.retrieve_offer_relations_query(offers_models.Offer.query)
-        .join(subquery, offers_models.Offer.id == subquery.c.max_id)
-        .all()
-    )
-
-
-def _serialize_products_from_body(
-    products: list[serialization.ProductOfferByEanCreation],
-) -> dict:
-    stock_details = {}
-    for product in products:
-        stock_details[product.ean] = {
-            "quantity": product.stock.quantity,
-            "price": product.stock.price,
-            "booking_limit_datetime": product.stock.booking_limit_datetime,
-        }
-    return stock_details
-
-
-def _create_offer_from_product(
-    venue: offerers_models.Venue,
-    product: offers_models.Product,
-    provider: providers_models.Provider,
-) -> offers_models.Offer:
-    ean = product.extraData.get("ean") if product.extraData else None
-
-    offer = offers_api.build_new_offer_from_product(venue, product, ean, provider.id)
-
-    offer.audioDisabilityCompliant = venue.audioDisabilityCompliant
-    offer.mentalDisabilityCompliant = venue.mentalDisabilityCompliant
-    offer.motorDisabilityCompliant = venue.motorDisabilityCompliant
-    offer.visualDisabilityCompliant = venue.visualDisabilityCompliant
-
-    offer.isActive = True
-    offer.lastValidationDate = datetime.datetime.utcnow()
-    offer.lastValidationType = OfferValidationType.AUTO
-    offer.lastValidationAuthorUserId = None
-
-    logger.info(
-        "models.Offer has been created",
-        extra={
-            "offer_id": offer.id,
-            "venue_id": venue.id,
-            "product_id": offer.productId,
-        },
-        technical_message_id="offer.created",
-    )
-
-    return offer
+    serialized_products_stocks = product_service.serialize_products_from_body(body.products)
+    product_service.create_or_update_ean_offers.delay(serialized_products_stocks, venue.id, current_api_key.provider.id)
 
 
 @blueprint.v1_offers_blueprint.route("/products/<int:product_id>", methods=["GET"])
@@ -672,46 +462,11 @@ def edit_product(body: serialization.ProductOfferEdition) -> serialization.Produ
             if body.image:
                 utils.save_image(body.image, updated_offer)
             if "stock" in updated_offer_from_body:
-                _upsert_product_stock(updated_offer, body.stock, current_api_key.provider)
+                product_service.upsert_product_stock(updated_offer, body.stock, current_api_key.provider)
     except (offers_exceptions.OfferCreationBaseException, offers_exceptions.OfferEditionBaseException) as e:
         raise api_errors.ApiErrors(e.errors, status_code=400)
 
     return serialization.ProductOfferResponse.build_product_offer(offer)
-
-
-def _upsert_product_stock(
-    offer: offers_models.Offer,
-    stock_body: serialization.StockEdition | None,
-    provider: providers_models.Provider,
-) -> None:
-    existing_stock = next((stock for stock in offer.activeStocks), None)
-    if not stock_body:
-        if existing_stock:
-            offers_api.delete_stock(existing_stock)
-        return
-
-    if not existing_stock:
-        if not stock_body.price:
-            raise api_errors.ApiErrors({"stock.price": ["Required"]})
-        offers_api.create_stock(
-            offer=offer,
-            price=finance_utils.to_euros(stock_body.price),
-            quantity=serialization.deserialize_quantity(stock_body.quantity),
-            booking_limit_datetime=stock_body.booking_limit_datetime,
-            creating_provider=provider,
-        )
-        return
-
-    stock_update_body = stock_body.dict(exclude_unset=True)
-    price = stock_update_body.get("price", offers_api.UNCHANGED)
-    quantity = serialization.deserialize_quantity(stock_update_body.get("quantity", offers_api.UNCHANGED))
-    offers_api.edit_stock(
-        existing_stock,
-        quantity=quantity + existing_stock.dnBookedQuantity if isinstance(quantity, int) else quantity,
-        price=finance_utils.to_euros(price) if price != offers_api.UNCHANGED else offers_api.UNCHANGED,
-        booking_limit_datetime=stock_update_body.get("booking_limit_datetime", offers_api.UNCHANGED),
-        editing_provider=provider,
-    )
 
 
 @blueprint.v1_offers_blueprint.route("/products/categories", methods=["GET"])
